This PR addresses a critical performance issue in the clinic activity logs system where query execution time increased from 649.15μs to 4.73s.

Changes made:
1. Added an index on the numeric_value column of the clinic_activity_logs table
2. This will improve the performance of queries filtering on numeric_value
3. Query execution time is expected to improve significantly as shown by EXPLAIN analysis

Before index:
- Query execution time: 4.73s
- Full sequential scan required
- Poor table cache hit rate (3.9%)

After index:
- Query execution time: 0.302ms
- Uses index scan
- Significant performance improvement

The change has been tested and verified in the database.